pyproject.toml
[project]
name = "merger-arb-agents"
version = "0.1.0"
description = "AI agent pipeline for merger arbitrage deal analysis"
requires-python = ">=3.11"

dependencies = [
    "anthropic>=0.40.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "tavily-python>=0.5.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "rich>=13.0.0",          # pretty terminal output during dev
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx",                  # for FastAPI test client
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/merger_arb"]
